corrected errors in readme, renamed robot_config and run_demo. for our mini robot, use *_mini.py files. The *_infinite_1.py is for coache's testing only.
