+++ .gitignore
@@ -1,0 +1,2 @@
+.vscode/launch.json
+.venv/** */
