--- example.py
@@ -1,49 +1,0 @@
-# pybricks blocks file:{"blocks":{"languageVersion":0,"blocks":[{"type":"blockGlobalSetup","id":"bjK,wS1MYO7aiYkFSwd{","x":0,"y":0,"deletable":false,"next":{"block":{"type":"variables_setup_imported_function","id":"Ps-i{bn2I6Ap)C]0Dx|d","fields":{"FROM_MODULE":{"id":"o@kKF~ZUh?W#x04[bOXQ"},"VAR":{"id":"RXs$DcB#pa@O`s,u5+gm"}},"next":{"block":{"type":"variables_setup_imported_function","id":";qeDD9EKS5j%MY=?)@Lq","fields":{"FROM_MODULE":{"id":"o@kKF~ZUh?W#x04[bOXQ"},"VAR":{"id":"G!$X]D.P~6c8uBK=)I:n"}},"next":{"block":{"type":"variables_setup_imported_function","id":";a,n%k[*|;L!6g$njWV)","fields":{"FROM_MODULE":{"id":"o@kKF~ZUh?W#x04[bOXQ"},"VAR":{"id":"2u-gn`JI8o$1m(C*J~wQ"}},"next":{"block":{"type":"variables_setup_imported_function","id":"i!MPJLXHnHuO?(7#XP7r","fields":{"FROM_MODULE":{"id":"o@kKF~ZUh?W#x04[bOXQ"},"VAR":{"id":"YEY8?aP0n)*s.9@8LBkS"}},"next":{"block":{"type":"variables_setup_imported_function","id":"OOWD8G]X~ahpHTJ9(@@T","fields":{"FROM_MODULE":{"id":"8Dk*.I8j)@u{;/kiTxHP"},"VAR":{"id":",pxLa(2E0oPEF,mi?KQ)"}},"next":{"block":{"type":"variables_setup_imported_function","id":"v,G/-7x*O19xz`5J=U=X","fields":{"FROM_MODULE":{"id":"OcnNg-Ikcu2*~eHNhbH3"},"VAR":{"id":"S.0:.ZE2gl-j8D_=~~)L"}},"next":{"block":{"type":"variables_setup_imported_function","id":"c?{Ra9|qYCiv^?VyfC1)","fields":{"FROM_MODULE":{"id":"OcnNg-Ikcu2*~eHNhbH3"},"VAR":{"id":"PT=es]hwInJ$0;(P0k~8"}}}}}}}}}}}}}}}}},{"type":"blockGlobalStart","id":"3tJe|AWl0baN(wH9a$@.","x":0,"y":407,"deletable":false,"next":{"block":{"type":"blockMultiTask","id":"!6nT^FfOh1pQQ%fp2h95","extraState":{"optionLevel":0},"fields":{"METHOD":"MULTITASK_ALL"},"inputs":{"TASK0":{"block":{"type":"blockComment","id":"v=]fRelaVW,0EIjX-Zo:","fields":{"FIELDNAME":"Blank,  done to make code multitask\nNeeded due to how the blocks work"}}}}}}},{"type":"variables_setup_function","id":"x1toQPx{:[:72QI1;hX1","x":0,"y":722,"extraState":{"optionLevel":0},"fields":{"ICON":"TASK","VAR":{"id":"G#O5gVezM6sna9uY|2ka"}},"inputs":{"STACK":{"block":{"type":"blockComment","id":"yAq9Jha?m!x*G+^XO`~R","fields":{"FIELDNAME":"Just a demo to show off the drivebase as well as music, icons,  \nand multitasking"},"next":{"block":{"type":"blockComment","id":"+r#dZOvyrocl{unn]gG3","fields":{"FIELDNAME":"Turn button yellow; and beep/wait so hands are out of the way"},"next":{"block":{"type":"blockLightOnColor","id":"iT{^7]d?,w.o*/hSk^UI","extraState":{"optionLevel":1},"fields":{"METHOD":"LIGHT_ON"},"inputs":{"VAR":{"shadow":{"type":"variables_get_color_light_device","id":"~m{Jz}bL3DDei_M=`w.[","fields":{"VAR":{"id":"KQyM,$VI4h{krVGQQc}Y","name":"HUB2","type":"PrimeHub"}}},"block":{"type":"blockreadOnlyGetter","id":"|I7P-DPTM[ozWDv2XwDJ","fields":{"VAR":{"id":"RXs$DcB#pa@O`s,u5+gm"}}}},"COLOR":{"shadow":{"type":"variables_get_color","id":"q=~2R%DmhgUbEK}$FBWh","fields":{"COLOUR":"#ffff00","VAR":{"id":"{-8bDu,EQr$dO+T[4=U`","name":"yellow","type":"ColorDef"}}}}},"next":{"block":{"type":"blockSpeakerBeep","id":"1Fqe.v//_E2PtJaz,_*q","inputs":{"VAR":{"shadow":{"type":"variables_get_speaker_hub","id":"J}`(sJDYNIrDuo2k3+dq","fields":{"VAR":{"id":"KQyM,$VI4h{krVGQQc}Y","name":"HUB2","type":"PrimeHub"}}},"block":{"type":"blockreadOnlyGetter","id":"3Zqn#fw@]7(3By;[7iCc","fields":{"VAR":{"id":"RXs$DcB#pa@O`s,u5+gm"}}}},"VALUE0":{"shadow":{"type":"unit_frequency","id":"/xlRy5pP_;{,sI$Hy%VD","fields":{"VALUE0":500}}},"VALUE1":{"shadow":{"type":"unit_time","id":"SjJUB-)0AY|ORtoSp,jf","fields":{"VALUE0":400}}}},"next":{"block":{"type":"blockMultiTask","id":"_xT2Qy~Ym9uFWz-(N:X2","extraState":{"optionLevel":0},"fields":{"METHOD":"MULTITASK_ALL"},"inputs":{"TASK0":{"block":{"type":"blockImportTaskCallStatement","id":"tJukK?UCQg@=Y_GDorfU","extraState":{"optionLevel":1},"inputs":{"VAR":{"shadow":{"type":"variables_get_imported_function","id":"{X9/Zq4}Igpw^k~_Y.;h","fields":{"VAR":{"id":"S.0:.ZE2gl-j8D_=~~)L","name":"display_pulse_icon","type":"FunctionImported"}}}},"ARG0":{"shadow":{"type":"blockMathNumber","id":"qMIG}e]t;W;wZ6FdxK~2","fields":{"NUM":0}},"block":{"type":"blockreadOnlyGetter","id":"(=nkC$tnZa:..c`hh{%9","fields":{"VAR":{"id":"PT=es]hwInJ$0;(P0k~8"}}}}},"next":{"block":{"type":"blockComment","id":"Tm|}FqED`]!p{Nb(h$OU","fields":{"FIELDNAME":"play some music to show off multitasking"},"next":{"block":{"type":"blockImportTaskCallStatement","id":"]f|$y-iOpeL_vOrzHM`n","extraState":{"optionLevel":0},"inputs":{"VAR":{"shadow":{"type":"variables_get_imported_function","id":"]atlEt-(iH,`6qiT({QC","fields":{"VAR":{"id":",pxLa(2E0oPEF,mi?KQ)","name":"star_wars_opening","type":"FunctionImported"}}}}}}}}}}},"TASK1":{"block":{"type":"blockComment","id":"g)qe+0T:cj5`(ZGKw^Ug","fields":{"FIELDNAME":"Turn on Gyro, drive forward"},"next":{"block":{"type":"blockDriveBaseUseGyro","id":"JRiY-R0v0MNEUYc4Pc]i","fields":{"METHOD":"DRIVEBASE_USE_GYRO_TRUE"},"inputs":{"VAR":{"shadow":{"type":"variables_get_drive_base_device","id":"D6O?{OF+f/Qv3*.TO{7w","fields":{"VAR":{"id":"J{tL/O{`|S{b(/:]n5HX","name":"DRIVE_BASE","type":"DriveBase"}}},"block":{"type":"blockreadOnlyGetter","id":"77D6:[9+5T:6xW=?hC[#","fields":{"VAR":{"id":"G!$X]D.P~6c8uBK=)I:n"}}}}},"next":{"block":{"type":"blockDriveBaseDrive","id":"U,utR8ZM(k/9X%v`8urV","extraState":{"optionLevel":2},"fields":{"METHOD":"DRIVEBASE_DRIVE_STRAIGHT"},"inputs":{"VAR":{"shadow":{"type":"variables_get_drive_base_device","id":"l6EPyh2Rss,jKgHeo2hU","fields":{"VAR":{"id":"J{tL/O{`|S{b(/:]n5HX","name":"DRIVE_BASE","type":"DriveBase"}}},"block":{"type":"blockreadOnlyGetter","id":"k4mZ9}xy3s}cH]6I-=pH","fields":{"VAR":{"id":"G!$X]D.P~6c8uBK=)I:n"}}}},"ARG0":{"shadow":{"type":"unit_distance","id":"jrr2]LV+3TNco5`XF##3","fields":{"VALUE0":100}}},"ARG1":{"shadow":{"type":"parameters_stop_4","id":"Gd6ISov%=:6x+XtKBmEX","fields":{"VALUE":"Stop.HOLD"}}}},"next":{"block":{"type":"blockFlowRepeat","id":"F#./$md7j]ZGN~~nn;Z[","inputs":{"TIMES":{"shadow":{"type":"blockMathNumber","id":"p6?n:~2B.tcr#5{g/F/:","fields":{"NUM":10}}},"DO":{"block":{"type":"blockDriveBaseDrive","id":"BMK#Wqe9E]qb7scBdg2c","extraState":{"optionLevel":2},"fields":{"METHOD":"DRIVEBASE_DRIVE_STRAIGHT"},"inputs":{"VAR":{"shadow":{"type":"variables_get_drive_base_device","id":"GqLb-N0tkP2?//kW*s^h","fields":{"VAR":{"id":"J{tL/O{`|S{b(/:]n5HX","name":"DRIVE_BASE","type":"DriveBase"}}},"block":{"type":"blockreadOnlyGetter","id":"s2L!g/II5~3II?IR0)me","fields":{"VAR":{"id":"G!$X]D.P~6c8uBK=)I:n"}}}},"ARG0":{"shadow":{"type":"unit_distance","id":"+evn71Yf/O3r@Gdrx--!","fields":{"VALUE0":300}}},"ARG1":{"shadow":{"type":"parameters_stop_4","id":"#.XGSfh9V|z_o$4ov{HF","fields":{"VALUE":"Stop.HOLD"}}}},"next":{"block":{"type":"blockDriveBaseDrive","id":"AmlIb|m00?D2cB2Wi?QK","extraState":{"optionLevel":3},"fields":{"METHOD":"DRIVEBASE_DRIVE_TURN"},"inputs":{"VAR":{"shadow":{"type":"variables_get_drive_base_device","id":"?hcZ},7tQct%{p3,s$iH","fields":{"VAR":{"id":"J{tL/O{`|S{b(/:]n5HX","name":"DRIVE_BASE","type":"DriveBase"}}},"block":{"type":"blockreadOnlyGetter","id":"fH=]%3`H.a~e-#Wi?69@","fields":{"VAR":{"id":"G!$X]D.P~6c8uBK=)I:n"}}}},"ARG0":{"shadow":{"type":"unit_angle","id":"?byz6*}iVLlrfEY|gV{)","fields":{"VALUE0":180}}},"ARG1":{"shadow":{"type":"parameters_stop_4","id":"fmY*J1z6/aT.7=iPR##K","fields":{"VALUE":"Stop.HOLD"}}}},"next":{"block":{"type":"blockDriveBaseDrive","id":"Sfr=1tjt_V6e3$T89MD9","extraState":{"optionLevel":2},"fields":{"METHOD":"DRIVEBASE_DRIVE_STRAIGHT"},"inputs":{"VAR":{"shadow":{"type":"variables_get_drive_base_device","id":"GqLb-N0tkP2?//kW*s^h","fields":{"VAR":{"id":"jD_|Nh!OfdSn;Y~=fn]O","name":"i","type":"DriveBase"}}},"block":{"type":"blockreadOnlyGetter","id":"RW+6ivQ/IaxPN*dGtchS","fields":{"VAR":{"id":"G!$X]D.P~6c8uBK=)I:n"}}}},"ARG0":{"shadow":{"type":"unit_distance","id":"|TmgLz1lQ-n,x#/f0m#G","fields":{"VALUE0":300}}},"ARG1":{"shadow":{"type":"parameters_stop_4","id":"u}hLh@`oqDa*S8[.lJ`.","fields":{"VALUE":"Stop.HOLD"}}}},"next":{"block":{"type":"blockDriveBaseDrive","id":"8S:{vLYyX#55Q?NS]erM","extraState":{"optionLevel":3},"fields":{"METHOD":"DRIVEBASE_DRIVE_TURN"},"inputs":{"VAR":{"shadow":{"type":"variables_get_drive_base_device","id":"?hcZ},7tQct%{p3,s$iH","fields":{"VAR":{"id":"jD_|Nh!OfdSn;Y~=fn]O","name":"i","type":"DriveBase"}}},"block":{"type":"blockreadOnlyGetter","id":"``~V7(QB^h+(NgCl*BTx","fields":{"VAR":{"id":"G!$X]D.P~6c8uBK=)I:n"}}}},"ARG0":{"shadow":{"type":"unit_angle","id":"7`%a1fhqp|Hn-mo!)3b_","fields":{"VALUE0":-180}}},"ARG1":{"shadow":{"type":"parameters_stop_4","id":"z4Wi$k.[f~k?OMePh:!I","fields":{"VALUE":"Stop.HOLD"}}}}}}}}}}}}},"next":{"block":{"type":"blockDriveBaseDrive","id":"VE~JKi`8Ea.rO%fT}.@9","extraState":{"optionLevel":2},"fields":{"METHOD":"DRIVEBASE_DRIVE_STRAIGHT"},"inputs":{"VAR":{"shadow":{"type":"variables_get_drive_base_device","id":"l6EPyh2Rss,jKgHeo2hU","fields":{"VAR":{"id":"jD_|Nh!OfdSn;Y~=fn]O","name":"i","type":"DriveBase"}}},"block":{"type":"blockreadOnlyGetter","id":"Js0_Me?_o?fgI0x;8=$C","fields":{"VAR":{"id":"G!$X]D.P~6c8uBK=)I:n"}}}},"ARG0":{"shadow":{"type":"unit_distance","id":"X~q!D]?%D{g2RBKY9L!D","fields":{"VALUE0":-100}}},"ARG1":{"shadow":{"type":"parameters_stop_4","id":"toRv+jLRd*O2tvIUb9w1","fields":{"VALUE":"Stop.HOLD"}}}},"next":{"block":{"type":"blockDriveBaseUseGyro","id":"#~NrO,@IC+hf,i];,:T~","fields":{"METHOD":"DRIVEBASE_USE_GYRO_FALSE"},"inputs":{"VAR":{"shadow":{"type":"variables_get_drive_base_device","id":"D6O?{OF+f/Qv3*.TO{7w","fields":{"VAR":{"id":"jD_|Nh!OfdSn;Y~=fn]O","name":"i","type":"DriveBase"}}},"block":{"type":"blockreadOnlyGetter","id":"9q=6;zv0a@2X!Z22v1SK","fields":{"VAR":{"id":"G!$X]D.P~6c8uBK=)I:n"}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}]},"variables":[{"name":"red","id":"fn;Nz9S21WEdU*:bePVZ","type":"ColorDef"},{"name":"orange","id":"mQT!K_m-~CHYjoeGBSgR","type":"ColorDef"},{"name":"yellow","id":"{-8bDu,EQr$dO+T[4=U`","type":"ColorDef"},{"name":"green","id":"yoKeQ5P,n2Yx!9=npfx~","type":"ColorDef"},{"name":"cyan","id":"-)N%)S9Qt%`CQ_p2K_Hs","type":"ColorDef"},{"name":"blue","id":"tMo1P/?Ifu,w+n4S+IyC","type":"ColorDef"},{"name":"violet","id":"G;}$OANB0*@Vb!:4]s$,","type":"ColorDef"},{"name":"magenta","id":"NQ?}TYKy=J5FnLilP:.s","type":"ColorDef"},{"name":"white","id":"[_y;WMd/OzKPgx,`rZV^","type":"ColorDef"},{"name":"none","id":";GL,gj,8X{GSF4}FtjSm","type":"ColorDef"},{"name":"run1","id":"G#O5gVezM6sna9uY|2ka","type":"Function"},{"name":"robot_config","id":"o@kKF~ZUh?W#x04[bOXQ","type":"ModuleImported"},{"name":"music_library","id":"8Dk*.I8j)@u{;/kiTxHP","type":"ModuleImported"},{"name":"icon_library","id":"OcnNg-Ikcu2*~eHNhbH3","type":"ModuleImported"},{"name":"HUB","id":"RXs$DcB#pa@O`s,u5+gm","type":"FunctionImported"},{"name":"RIGHT_ATTACHMENT","id":"YEY8?aP0n)*s.9@8LBkS","type":"FunctionImported"},{"name":"LEFT_ATTACHMENT","id":"2u-gn`JI8o$1m(C*J~wQ","type":"FunctionImported"},{"name":"DRIVE_BASE","id":"G!$X]D.P~6c8uBK=)I:n","type":"FunctionImported"},{"name":"star_wars_opening","id":",pxLa(2E0oPEF,mi?KQ)","type":"FunctionImported"},{"name":"display_pulse_icon","id":"S.0:.ZE2gl-j8D_=~~)L","type":"FunctionImported"},{"name":"REBEL","id":"PT=es]hwInJ$0;(P0k~8","type":"FunctionImported"}],"info":{"type":"pybricks","version":"1.3.2"}}
-from pybricks.parameters import Color
-from pybricks.tools import multitask, run_task, wait
-
-from icon_library import REBEL, display_pulse_icon
-from music_library import star_wars_opening
-from robot_config import DRIVE_BASE, HUB#, LEFT_ATTACHMENT, RIGHT_ATTACHMENT
-
-async def subtask():
-    await display_pulse_icon(REBEL)
-    # play some music to show off multitasking
-    await star_wars_opening()
-
-async def subtask2():
-    # Turn on Gyro, drive forward
-    DRIVE_BASE.use_gyro(True)
-    #await DRIVE_BASE.straight(100) #in mm
-    for count in range(1):
-        await wait(1000)
-        await DRIVE_BASE.straight(50)
-        await DRIVE_BASE.turn(180)
-        await wait(1000)
-        await DRIVE_BASE.straight(-50)
-        await DRIVE_BASE.turn(-180)
-    #await DRIVE_BASE.straight(-10)
-    DRIVE_BASE.use_gyro(False)
-
-async def run1():
-    await wait(0)
-    # Just a demo to show off the drivebase as well as music, icons,
-    # and multitasking
-    # Turn button yellow; and beep/wait so hands are out of the way
-    HUB.light.on(Color.YELLOW)
-    await HUB.speaker.beep(500, 400)
-    await multitask(
-        subtask(),
-        subtask2(),
-    )
-
-async def main():
-    # Blank,  done to make code multitask
-    # Needed due to how the blocks work
-    await multitask(
-        wait(0),
-        run1()
-    )
-
-
-run_task(main())
