--- robot_config_bakup.py
@@ -1,34 +1,0 @@
-# pybricks blocks file:{"blocks":{"languageVersion":0,"blocks":[{"type":"blockGlobalSetup","id":"bjK,wS1MYO7aiYkFSwd{","x":0,"y":0,"deletable":false,"next":{"block":{"type":"variables_set_prime_hub","id":"+u*2%4U=;qXA?zJ,AQfu","extraState":{"optionLevel":1},"fields":{"VAR":{"id":"v`MM3[Di(y|,tU9OY(Ym"}},"inputs":{"AXIS_TOP":{"shadow":{"type":"blockParametersAxis","id":"LHL[c}2a/1p$U;CbEzly","fields":{"VALUE":"z"}}},"AXIS_FRONT":{"shadow":{"type":"blockParametersAxis","id":"?4fbdDg%j9y~00Id7JBz","fields":{"VALUE":"x"}}}},"next":{"block":{"type":"variables_set_motor","id":"~.H.Tgx$i:^qUp/WqT0%","fields":{"VAR":{"id":"^_wvb!Nt$$MNi2fnY8AL"}},"inputs":{"PORT":{"shadow":{"type":"blockParametersPort","id":"Yz#/|BcaK;cS]5/w9*::","fields":{"NAME":"E"}}},"POSITIVE_DIRECTION":{"shadow":{"type":"blockParametersDirection","id":"?J]%{d.X;Em:A?0@At66","fields":{"SELECTION":"Direction.COUNTERCLOCKWISE"}}}},"next":{"block":{"type":"variables_set_motor","id":"uR,,fOKEhZ7!Jif24eWG","fields":{"VAR":{"id":"C2O6|U2:c(6D@Bwalggw"}},"inputs":{"PORT":{"shadow":{"type":"blockParametersPort","id":"9m9?V)f;Ki/KkIu]4IV=","fields":{"NAME":"F"}}},"POSITIVE_DIRECTION":{"shadow":{"type":"blockParametersDirection","id":".6T^}R^%d,wC-nKWnC0S","fields":{"SELECTION":"Direction.CLOCKWISE"}}}},"next":{"block":{"type":"variables_set_drive_base","id":"/vvbo(#N.Nds5!a?$/h$","fields":{"VAR":{"id":"4OHjWk}_Zbrg~SrAEUu%"}},"inputs":{"VAR":{"shadow":{"type":"variables_get_motor_device","id":"AxYx_ok)`0Ni$?aY9B={","fields":{"VAR":{"id":"^_wvb!Nt$$MNi2fnY8AL","name":"DRIVE_LEFT","type":"Motor"}}}},"VAR2":{"shadow":{"type":"variables_get_motor_device","id":"QACX0t4^9TDNy;[q~zEw","fields":{"VAR":{"id":"C2O6|U2:c(6D@Bwalggw","name":"DRIVE_RIGHT","type":"Motor"}}}},"VALUE0":{"shadow":{"type":"unit_distance","id":"xZ2?NXxVwo7`-y7*?(%V","fields":{"VALUE0":88}}},"VALUE1":{"shadow":{"type":"unit_distance","id":"Dp`EM+O-v0NPCfsMOaBx","fields":{"VALUE0":128}}}},"next":{"block":{"type":"variables_set_motor","id":"w0!A_@PIt2RdsK8(/Hbl","fields":{"VAR":{"id":"?nUAjPA?`Axz*{IVIYH^"}},"inputs":{"PORT":{"shadow":{"type":"blockParametersPort","id":"8bz=cRVl3@F~o/*pdI%7","fields":{"NAME":"C"}}},"POSITIVE_DIRECTION":{"shadow":{"type":"blockParametersDirection","id":"AL`VON|y6U~[5B5SH_?~","fields":{"SELECTION":"Direction.CLOCKWISE"}}}},"next":{"block":{"type":"variables_set_motor","id":"[hA@Ll{Y^+Rqq?hXuJki","fields":{"VAR":{"id":"$Ujk8G70lPWj*bY@8^Qc"}},"inputs":{"PORT":{"shadow":{"type":"blockParametersPort","id":"!}XzCM+pVq4D9`(tP[JQ","fields":{"NAME":"D"}}},"POSITIVE_DIRECTION":{"shadow":{"type":"blockParametersDirection","id":"]9L=a]BH1Gp?YDpvHrEv","fields":{"SELECTION":"Direction.CLOCKWISE"}}}},"next":{"block":{"type":"variables_set_color_sensor","id":"xp?xypLF:*|PeZK},hk+","extraState":{"optionLevel":0},"fields":{"VAR":{"id":"A)}jFd)]X][j2D[=.Q?V"}},"inputs":{"PORT":{"shadow":{"type":"blockParametersPort","id":"Mz_j=SqH_jKY#bXTe:iD","fields":{"NAME":"A"}}}},"next":{"block":{"type":"variables_set_color_sensor","id":"lmDi-e2E^;)LELn0uw/I","extraState":{"optionLevel":0},"fields":{"VAR":{"id":"u8rA(KDl|9cYI1;,Kn0d"}},"inputs":{"PORT":{"shadow":{"type":"blockParametersPort","id":"~Q-.Wl!rEEa9=.;fc^./","fields":{"NAME":"B"}}}},"next":{"block":{"type":"variables_setup_any","id":"AW20{#;{)8d$I*hy$bkV","fields":{"VAR":{"id":"Gw*-]G.*#g+:u|OyS6/Q"}},"inputs":{"VALUE0":{"shadow":{"type":"blockMathNumber","id":"toxIC]AAHUn;5Q.A95$l","fields":{"NUM":620}}}},"next":{"block":{"type":"variables_setup_any","id":"$$S?g1{6~3r4Xhm.aTN=","fields":{"VAR":{"id":"?V):uCt{BesFh%[!/{vg"}},"inputs":{"VALUE0":{"shadow":{"type":"blockMathNumber","id":"{g7+{Sf5XDLZCXV~Iz__","fields":{"NUM":600}}}},"next":{"block":{"type":"variables_setup_any","id":"?D7v6eV@BP.:|,5bV09U","fields":{"VAR":{"id":"nAEvOKX}[FxeFHOW/qHW"}},"inputs":{"VALUE0":{"shadow":{"type":"blockMathNumber","id":"eVM+9@NG4MU^OF#_zXLW","fields":{"NUM":400}}}},"next":{"block":{"type":"variables_setup_any","id":"!lj$DB.NXM{+#T.~/}zy","fields":{"VAR":{"id":"Q};0~!05]:fH;AnO2].("}},"inputs":{"VALUE0":{"shadow":{"type":"blockMathNumber","id":"PE=gtRrVvV?OPdUw6p_%","fields":{"NUM":1000}}}},"next":{"block":{"type":"variables_setup_any","id":"D@n9i0~0=1py+N`^^0G(","fields":{"VAR":{"id":"f~]mZQP0#S4Pmnj](@S0"}},"inputs":{"VALUE0":{"shadow":{"type":"blockMathNumber","id":"p*y1Eo{dsE:[7!:Nx9$5","fields":{"NUM":250}}}},"next":{"block":{"type":"variables_setup_any","id":"OS2@7Bb9I8Nr?K^Ip2*Y","fields":{"VAR":{"id":"=dPf$nwOv588oY{TJsT}"}},"inputs":{"VALUE0":{"shadow":{"type":"blockMathNumber","id":"#4BTXKK:sKMQCw1l6NQg","fields":{"NUM":300}}}},"next":{"block":{"type":"variables_setup_any","id":"E|T408GmBSIZfG|-,Tk%","fields":{"VAR":{"id":"(!m3~{5_%i!=i].n4L_b"}},"inputs":{"VALUE0":{"shadow":{"type":"blockMathNumber","id":"2am)oHxBiA3)6oY6rH0G","fields":{"NUM":1000}}}},"next":{"block":{"type":"variables_setup_any","id":"vt[ifIo|Ye6fA%3$ZB$i","fields":{"VAR":{"id":"M?-q[$D}uH]nx$oVM?J+"}},"inputs":{"VALUE0":{"shadow":{"type":"blockMathNumber","id":"bQ2u1py[s?`cvsj@9LV6","fields":{"NUM":200}}}},"next":{"block":{"type":"variables_setup_any","id":"|UzgdY}i*Xr?=CQp|_6L","fields":{"VAR":{"id":"R`RK-U0#Ug/JUTsF~6*="}},"inputs":{"VALUE0":{"shadow":{"type":"blockMathNumber","id":"0QpkQ.7C?,$D?DS#G=qz","fields":{"NUM":1000}}}},"next":{"block":{"type":"variables_setup_any","id":".|frV2?h*o%oCbIVPCQ1","fields":{"VAR":{"id":"FXTZvX$bR6WtO~]#Hh@8"}},"inputs":{"VALUE0":{"shadow":{"type":"blockMathNumber","id":"%^k~irhuw?SmRj2K?yR7","fields":{"NUM":600}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}},{"type":"blockGlobalStart","id":"3tJe|AWl0baN(wH9a$@.","x":0,"y":925,"deletable":false,"next":{"block":{"type":"blockComment","id":"Lo]C5E|qQb1?d/`u)4;)","fields":{"FIELDNAME":"This is where you configure all the variables as well as\nthe parts of the robot. These values are then imported/used as necessary.\n\nChange to match the setup of your robot. Adjust speeds as necessary"}}}}]},"variables":[{"name":"red","id":"iH)Hpm#0yY8Oe);dmGQL","type":"ColorDef"},{"name":"orange","id":"2:^klv!#}GXk%=S|Dt*^","type":"ColorDef"},{"name":"yellow","id":"6je/;ktPG7-.AxMuk9So","type":"ColorDef"},{"name":"green","id":"r-c?0%0E9ck!(Op9b4Lo","type":"ColorDef"},{"name":"cyan","id":"P[zZh~o8BDoc~7n#kNQc","type":"ColorDef"},{"name":"blue","id":"XUoM?u7L76C5dt5oj_Oi","type":"ColorDef"},{"name":"violet","id":"-f857pmudyvx07;gaug@","type":"ColorDef"},{"name":"magenta","id":"*Mc;CIhVmDTQ(MpZRfrM","type":"ColorDef"},{"name":"white","id":"2^`7+$#[(;r|-^`S[C#a","type":"ColorDef"},{"name":"none","id":"zm1a|%B_jQ{su%H`n_Em","type":"ColorDef"},{"name":"LEFT_COLOR_SENSOR","id":"A)}jFd)]X][j2D[=.Q?V","type":"ColorSensor"},{"name":"RIGHT_COLOR_SENSOR","id":"u8rA(KDl|9cYI1;,Kn0d","type":"ColorSensor"},{"name":"DRIVE_BASE","id":"4OHjWk}_Zbrg~SrAEUu%","type":"DriveBase"},{"name":"RIGHT_ATTACHMENT","id":"$Ujk8G70lPWj*bY@8^Qc","type":"Motor"},{"name":"LEFT_ATTACHMENT","id":"?nUAjPA?`Axz*{IVIYH^","type":"Motor"},{"name":"DRIVE_RIGHT","id":"C2O6|U2:c(6D@Bwalggw","type":"Motor"},{"name":"DRIVE_LEFT","id":"^_wvb!Nt$$MNi2fnY8AL","type":"Motor"},{"name":"HUB","id":"v`MM3[Di(y|,tU9OY(Ym","type":"PrimeHub"},{"name":"SPEED","id":"Gw*-]G.*#g+:u|OyS6/Q","type":"Any"},{"name":"ACCELERATION","id":"?V):uCt{BesFh%[!/{vg","type":"Any"},{"name":"TURN_SPEED","id":"nAEvOKX}[FxeFHOW/qHW","type":"Any"},{"name":"TURN_ACCELERATION","id":"Q};0~!05]:fH;AnO2].(","type":"Any"},{"name":"MANUAL_MOTOR_SPEED","id":"f~]mZQP0#S4Pmnj](@S0","type":"Any"},{"name":"TELEOP_SPEED","id":"=dPf$nwOv588oY{TJsT}","type":"Any"},{"name":"TELEOP_TURN_ACCEL","id":"R`RK-U0#Ug/JUTsF~6*=","type":"Any"},{"name":"TELEOP_TURN","id":"M?-q[$D}uH]nx$oVM?J+","type":"Any"},{"name":"TELEOP_ACCEL","id":"(!m3~{5_%i!=i].n4L_b","type":"Any"},{"name":"TELEOP_ATTACH_SPEED","id":"FXTZvX$bR6WtO~]#Hh@8","type":"Any"}],"info":{"type":"pybricks","version":"1.3.2"}}
-from pybricks.hubs import PrimeHub
-from pybricks.parameters import Axis, Direction, Port
-from pybricks.pupdevices import ColorSensor, Motor
-from pybricks.robotics import DriveBase
-
-# Set up all devices.
-HUB = PrimeHub(top_side=Axis.Z, front_side=Axis.X)
-LEFT_COLOR_SENSOR = ColorSensor(Port.A)
-RIGHT_COLOR_SENSOR = ColorSensor(Port.B)
-DRIVE_LEFT = Motor(Port.E, Direction.COUNTERCLOCKWISE)
-DRIVE_RIGHT = Motor(Port.F, Direction.CLOCKWISE)
-DRIVE_BASE = DriveBase(DRIVE_LEFT, DRIVE_RIGHT, 88, 128)
-LEFT_ATTACHMENT = Motor(Port.C, Direction.CLOCKWISE)
-RIGHT_ATTACHMENT = Motor(Port.D, Direction.CLOCKWISE)
-
-# Initialize variables.
-SPEED = 620
-ACCELERATION = 600
-TURN_SPEED = 400
-TURN_ACCELERATION = 1000
-MANUAL_MOTOR_SPEED = 250
-TELEOP_SPEED = 300
-TELEOP_ACCEL = 1000
-TELEOP_TURN = 200
-TELEOP_TURN_ACCEL = 1000
-TELEOP_ATTACH_SPEED = 600
-
-
-# The main program starts here.
-# This is where you configure all the variables as well as
-# the parts of the robot. These values are then imported/used as necessary.
-
-# Change to match the setup of your robot. Adjust speeds as necessary
